+++ src/main/scala/com/mogproject/mogami/frontend/sam/SAMObserver.scala
@@ -5,16 +5,16 @@ package com.mogproject.mogami.frontend.sam
   */
 trait SAMObserver[M <: SAMModel] {
 
-  def samObserveMask: Int
+  def samObserveMask(): Long
 
   /**
     *
     * @param model model
     * @param flag  -1: all bits on => refresh all
     */
-  def refresh(model: M, flag: Int): Unit
+  def refresh(model: M, flag: Long): Unit
 
-  protected def isFlagUpdated(flag: Int, mask: Int): Boolean = (flag & mask) != 0
+  protected def isFlagUpdated(flag: Long, mask: Long): Boolean = (flag & mask) != 0
 
   SAM.addObserver(this)
